pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ruminant"
description = "A CLI tool for tracking activity across community projects"
readme = "README.md"
license = { file = "LICENSE.md" }
authors = [{ name = "Anil Madhavapeddy", email = "anil@recoil.org" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.10"
dependencies = [
    "typer[rich]>=0.9.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
    "tomli>=2.0.1",
    "tomli-w>=1.0.0",
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/avsm/ruminant"
Repository = "https://github.com/avsm/ruminant.git"
Issues = "https://github.com/avsm/ruminant/issues"

[project.scripts]
ruminant = "ruminant.main:app"

[tool.hatch.version]
path = "ruminant/__init__.py"

[tool.hatch.build.targets.sdist]
include = [
    "/ruminant",
    "/README.md",
    "/LICENSE",
]

[tool.hatch.build.targets.wheel]
packages = ["ruminant"]
